Move SG code into separate process
