+++ src/modules/ShortcutGuide/ShortcutGuide/LocProject.json
@@ -1,0 +1,14 @@
+{
+    "Projects": [
+        {
+            "LanguageSet": "Azure_Languages",
+            "LocItems": [
+                {
+                    "SourceFile": "src\\modules\\shortcut_guide\\Resources.resx",
+                    "CopyOption": "LangIDOnName",
+                    "OutputPath": "src\\modules\\shortcut_guide"
+                }
+            ]
+        }
+    ]
+}
