+++ src/modules/ShortcutGuide/ShortcutGuide/ShortcutGuide.vcxproj.filters
@@ -13,9 +13,51 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
+    <Filter Include="svgs">
+      <UniqueIdentifier>{800507f8-718d-48d1-aa56-96c040795b8a}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="animation.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="d2d_svg.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="d2d_text.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="d2d_window.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="keyboard_state.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="native_event_waiter.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="overlay_window.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="shortcut_guide.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ShortcutGuideConstants.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="start_visible.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="target_state.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="tasklist_positions.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="trace.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -26,12 +68,86 @@
     <ClCompile Include="main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="animation.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="d2d_svg.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="d2d_text.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="d2d_window.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="keyboard_state.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="native_event_waiter.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="overlay_window.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="shortcut_guide.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="start_visible.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="target_state.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="tasklist_positions.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="trace.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="PropertySheet.props" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Text Include="readme.txt" />
+    <CopyFileToFolders Include="svgs\0.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\1.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\2.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\3.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\4.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\5.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\6.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\7.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\8.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\9.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\no_active_window.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\overlay.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
+    <CopyFileToFolders Include="svgs\overlay_portrait.svg">
+      <Filter>svgs</Filter>
+    </CopyFileToFolders>
   </ItemGroup>
 </Project>
