+++ src/modules/ShortcutGuide/ShortcutGuide/animation.cpp
