+++ src/modules/ShortcutGuide/ShortcutGuide/animation.h
