+++ src/modules/ShortcutGuide/ShortcutGuide/d2d_svg.h
