+++ src/modules/ShortcutGuide/ShortcutGuide/d2d_text.h
