+++ src/modules/ShortcutGuide/ShortcutGuide/d2d_window.h
