+++ src/modules/ShortcutGuide/ShortcutGuide/pch.cpp
@@ -1,1 +1,1 @@
-﻿#include "pch.h"
+#include "pch.h"
