+++ src/modules/ShortcutGuide/ShortcutGuide/resource.base.h
@@ -1,0 +1,13 @@
+//{{NO_DEPENDENCIES}}
+// Microsoft Visual C++ generated include file.
+// Used by ShortcutGuide.rc
+
+//////////////////////////////
+// Non-localizable
+
+#define FILE_DESCRIPTION "PowerToys ShortcutGuide"
+#define INTERNAL_NAME "ShortcutGuide"
+#define ORIGINAL_FILENAME "ShortcutGuide.dll"
+
+// Non-localizable
+//////////////////////////////
