+++ src/modules/ShortcutGuide/ShortcutGuideModuleInterface/ShortcutGuideModuleInterface.vcxproj
@@ -11,7 +11,6 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
-  <!-- todo: reconsider -->
   <Target Name="GenerateResourceFiles" BeforeTargets="PrepareForBuild">
     <Exec Command="powershell -NonInteractive -executionpolicy Unrestricted $(SolutionDir)tools\build\convert-resx-to-rc.ps1 $(MSBuildThisFileDirectory) resource.base.h resource.h ShortcutGuide.base.rc ShortcutGuide.rc" />
   </Target>
@@ -64,61 +63,32 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="animation.h" />
-    <ClInclude Include="d2d_svg.h" />
-    <ClInclude Include="d2d_text.h" />
-    <ClInclude Include="d2d_window.h" />
-    <ClInclude Include="framework.h" />
-    <ClInclude Include="keyboard_state.h" />
-    <ClInclude Include="native_event_waiter.h" />
-    <ClInclude Include="overlay_window.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="resource.base.h" />
-    <ClInclude Include="resource.h" />
-    <ClInclude Include="ShortcutGuideConstants.h" />
     <ClInclude Include="shortcut_guide.h" />
-    <ClInclude Include="start_visible.h" />
-    <ClInclude Include="target_state.h" />
-    <ClInclude Include="tasklist_positions.h" />
     <ClInclude Include="trace.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="animation.cpp" />
-    <ClCompile Include="d2d_svg.cpp" />
-    <ClCompile Include="d2d_text.cpp" />
-    <ClCompile Include="d2d_window.cpp" />
     <ClCompile Include="dllmain.cpp" />
-    <ClCompile Include="keyboard_state.cpp" />
-    <ClCompile Include="native_event_waiter.cpp" />
-    <ClCompile Include="overlay_window.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(CIBuild)'!='true'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="shortcut_guide.cpp" />
-    <ClCompile Include="start_visible.cpp" />
-    <ClCompile Include="target_state.cpp" />
-    <ClCompile Include="tasklist_positions.cpp" />
     <ClCompile Include="trace.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="ShortcutGuide.base.rc" />
+    <ProjectReference Include="..\..\..\common\logger\logger.vcxproj">
+      <Project>{d9b8fc84-322a-4f9f-bbb9-20915c47ddfd}</Project>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources.resx" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\..\common\Display\Display.vcxproj">
-      <Project>{caba8dfb-823b-4bf2-93ac-3f31984150d9}</Project>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\common\logger\logger.vcxproj">
-      <Project>{d9b8fc84-322a-4f9f-bbb9-20915c47ddfd}</Project>
-    </ProjectReference>
-    <ProjectReference Include="..\..\..\common\Themes\Themes.vcxproj">
-      <Project>{98537082-0fdb-40de-abd8-0dc5a4269bab}</Project>
-    </ProjectReference>
+    <ResourceCompile Include="ShortcutGuide.base.rc" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <Import Project="..\..\..\..\deps\spdlog.props" />
