+++ src/modules/ShortcutGuide/ShortcutGuideModuleInterface/dllmain.cpp
@@ -2,7 +2,6 @@
 #include "pch.h"
 #include <mutex>
 #include "shortcut_guide.h"
-#include "overlay_window.h"
 #include "trace.h"
 
 BOOL APIENTRY DllMain(HMODULE hModule, DWORD ul_reason_for_call, LPVOID lpReserved)
