+++ src/modules/ShortcutGuide/ShortcutGuideModuleInterface/pch.h
@@ -1,29 +1,8 @@
 #pragma once
 #define NOMINMAX
 #include <winrt/base.h>
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Foundation.Collections.h>
 #include <Windows.h>
-#include <dxgi1_3.h>
-#include <d3d11_2.h>
-#include <d2d1_3.h>
-#include <d2d1_3helper.h>
-#include <d2d1helper.h>
-#include <dwrite.h>
-#include <dcomp.h>
-#include <dwmapi.h>
-#include <Shobjidl.h>
-#include <Shlwapi.h>
-#include <string>
-#include <algorithm>
-#include <chrono>
-#include <mutex>
-#include <thread>
-#include <functional>
-#include <condition_variable>
-#include <stdexcept>
-#include <tuple>
-#include <unordered_set>
-#include <string>
 #include <ProjectTelemetry.h>
 #include <filesystem>
+#include <common/logger/logger.h>
+#include <common/utils/resources.h>
