+++ src/modules/ShortcutGuide/ShortcutGuideModuleInterface/trace.cpp
@@ -18,30 +18,6 @@ void Trace::UnregisterProvider() noexcept
     TraceLoggingUnregister(g_hProvider);
 }
 
-void Trace::HideGuide(const __int64 duration_ms, std::vector<int>& key_pressed) noexcept
-{
-    std::string vk_codes;
-    std::vector<int>::iterator it;
-    for (it = key_pressed.begin(); it != key_pressed.end();)
-    {
-        vk_codes += std::to_string(*it);
-        if (++it != key_pressed.end())
-        {
-            vk_codes += " ";
-        }
-    }
-
-    TraceLoggingWrite(
-        g_hProvider,
-        "ShortcutGuide_HideGuide",
-        TraceLoggingInt64(duration_ms, "DurationInMs"),
-        TraceLoggingInt64(key_pressed.size(), "NumberOfKeysPressed"),
-        TraceLoggingString(vk_codes.c_str(), "ListOfKeysPressed"),
-        ProjectTelemetryPrivacyDataTag(ProjectTelemetryTag_ProductAndServicePerformance),
-        TraceLoggingBoolean(TRUE, "UTCReplace_AppSessionGuid"),
-        TraceLoggingKeyword(PROJECT_KEYWORD_MEASURE));
-}
-
 void Trace::EnableShortcutGuide(const bool enabled) noexcept
 {
     TraceLoggingWrite(
@@ -51,30 +27,4 @@ void Trace::EnableShortcutGuide(const bool enabled) noexcept
         ProjectTelemetryPrivacyDataTag(ProjectTelemetryTag_ProductAndServicePerformance),
         TraceLoggingBoolean(TRUE, "UTCReplace_AppSessionGuid"),
         TraceLoggingKeyword(PROJECT_KEYWORD_MEASURE));
-}
-
-void Trace::SettingsChanged(const int press_delay_time, const int overlay_opacity, const std::wstring& theme) noexcept
-{
-    TraceLoggingWrite(
-        g_hProvider,
-        "ShortcutGuide_SettingsChanged",
-        TraceLoggingInt32(press_delay_time, "PressDelayTime"),
-        TraceLoggingInt32(overlay_opacity, "OverlayOpacity"),
-        TraceLoggingWideString(theme.c_str(), "Theme"),
-        ProjectTelemetryPrivacyDataTag(ProjectTelemetryTag_ProductAndServicePerformance),
-        TraceLoggingBoolean(TRUE, "UTCReplace_AppSessionGuid"),
-        TraceLoggingKeyword(PROJECT_KEYWORD_MEASURE));
-}
-
-// Log if an error occurs in Shortcut Guide
-void Trace::Error(const DWORD errorCode, std::wstring errorMessage, std::wstring methodName) noexcept
-{
-    TraceLoggingWrite(
-        g_hProvider,
-        "ShortcutGuide_Error",
-        ProjectTelemetryPrivacyDataTag(ProjectTelemetryTag_ProductAndServicePerformance),
-        TraceLoggingKeyword(PROJECT_KEYWORD_MEASURE),
-        TraceLoggingValue(methodName.c_str(), "MethodName"),
-        TraceLoggingValue(errorCode, "ErrorCode"),
-        TraceLoggingValue(errorMessage.c_str(), "ErrorMessage"));
 }
