+++ src/runner/runner.vcxproj
@@ -86,71 +86,6 @@
     <ResourceCompile Include="Generated Files/runner.rc" />
   </ItemGroup>
   <ItemGroup>
-    <CopyFileToFolders Include="svgs\0.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\1.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\2.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\3.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\4.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\5.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\6.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\7.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\8.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\9.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\no_active_window.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\overlay.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\overlay_portrait.svg">
-      <DeploymentContent>true</DeploymentContent>
-      <FileType>Document</FileType>
-      <DestinationFolders>$(OutDir)\svgs</DestinationFolders>
-    </CopyFileToFolders>
     <CopyFileToFolders Include="svgs\icon.ico">
       <DeploymentContent>true</DeploymentContent>
       <FileType>Document</FileType>
