+++ src/runner/runner.vcxproj.filters
@@ -95,9 +95,6 @@
     <Filter Include="Utils">
       <UniqueIdentifier>{bc27e9c1-8afa-4d62-8179-789f4651c0b6}</UniqueIdentifier>
     </Filter>
-    <Filter Include="svgs">
-      <UniqueIdentifier>{1123565f-43ba-4eb8-a6f5-2d5f8f3ae6af}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Resource Files">
       <UniqueIdentifier>{74d0b535-16ee-46cc-9adf-dc4668bbcfda}</UniqueIdentifier>
     </Filter>
@@ -107,45 +104,6 @@
   </ItemGroup>
   <ItemGroup>
     <CopyFileToFolders Include="svgs\icon.ico" />
-    <CopyFileToFolders Include="svgs\0.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\1.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\2.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\3.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\4.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\5.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\6.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\7.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\8.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\9.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\no_active_window.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\overlay.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
-    <CopyFileToFolders Include="svgs\overlay_portrait.svg">
-      <Filter>svgs</Filter>
-    </CopyFileToFolders>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
